외부 API 연동 & 테스트 로직 작성

+++ src/main/java/io/github/seoleeder/owls_pick/entity/game/GameCompanyId.java
@@ -12,6 +12,6 @@ import java.io.Serializable;
 @EqualsAndHashCode // 복합키 식별을 위해 필수
 public class GameCompanyId implements Serializable {
 
-    private int gameId;
-    private int companyId;
+    private Long gameId;
+    private Long companyId;
 }
